Update the widgets to be containers - #4683
